Code improved but still needing testing
A slow clock is now being generated to send to the controllers instead
of the 100 MHz one which is unsupported by HFE4021B. 200 kHz was chosen
and the reason is to avoid delays and noises.

The clock is also inverted before it's sent to the controller (to avoid
delays too) and so the Basys is half period ahead of the controller to
be ready to send and receive data before the controller reads or sends
its data.

The time ratio to write and read the controller data is now on 50%.
Probably it's too much because the software runs much faster than the
slow clock.

More macros used to make the code less error prone and and sequential
and combinational logic separated for easier debugging.