Controller code ready to work
Software and hardware data registers size changed to 8 bits. The write
operation on the registers is done in a way that doesn't require an
additional bit to say if the operation is complete or not.

Code that orders and receives the controller data is now fully
implemented and works. The data is read every 10 slow_clk cycles.

Also macros are now being used instead of manual implementations of
registers and counters (much easier and less error prone).

And at least currently it's needed to have 2 outputs to give to each
controller, for some reason, because if one output (for example "pl") is
used for both controllers, one of them doesn't receive the signal (but
this needs testing after removing all slow_clk usages except on the
output).

And removed the latches that were created on the combinatory logic
block - replaced by registers.